Initial commit of a working web server
This commit introduces a functional web server. Using the handler package handlers are registered as a NewWebServer is created if those handlers support HTTP. This commit also introduces the ability to render web pages using the template package and provides support for a layout and view behavior. Finally, this commit also introduces the concept of a RequestEvent which will be used to manage details about a specific request and feed into O.C.E.AN.

The expected result of this commit is that the application which includes this library can compile and support basic web request.

* Added a functional web server
* Added RequestEvent
* Added ability to render layouts and views using templates
* Added a working example of tests using the ginkgo package which supports BDD style tests.