WBOT-2 - Routing, Handler Execution, and Logging
Logging
The Web server now uses the standard library's logging package to efficiently log to the stdout.

Server Factory
The New function now sets some settings on the render package. This commit simply serves to establish the process.

Handler Execution
A single handler can now be registered and executed as expected. Soon multiple handlers will be able to be executed in a chain for a matched route.As a route is matched the request is captured generating an Event which is provided to handlers during execution.