WBOT-2 - 404 Response and Header Improvement
The webserver now attempts to render a configured template when a handler can not be matched aginst the requested route. If the configured template for `onMissingHandler` can not be rendered an optional error will be logged and the system will render a backup 404 response.

This update also standardizes the debug and error logging settings.