Context JSON removed; Context StatusCode removed
The StatusCode property of the context is not to be used and has been removed. This was a relect from before the creation fo input and output.

This commit also removes the JSON tagging for the context object. As of now there is no reason to marshal this object into JSON.